+++ RssReader/Controller/Setting/SettingViewController.swift
@@ -1,0 +1,23 @@
+//
+//  SettingViewController.swift
+//  RssReader
+//
+//  Created by 若江照仁 on 2021/02/17.
+//
+
+import UIKit
+import Nuke
+
+class SettingViewController: UIViewController {
+    //MARK:- @IBOutlet
+    
+    @IBOutlet weak var settingTableView: UITableView!
+    
+    //MARK:- 変数宣言
+    
+    //MARK:- ライフサイクル関連
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+    }
+}
